+++ console/src/components/Services/Data/TableBrowseRows/utils.ts
@@ -1,3 +1,5 @@
+import { Operators } from '../constants';
+
 type TableSchema = {
   primary_key?: { columns: string[] };
   columns: Array<{ column_name: string }>;
@@ -36,6 +38,54 @@ export const compareRows = (
     }
   });
   return same;
+};
+
+export const getDefaultValue = (possibleValue: unknown, opName: string) => {
+  if (possibleValue) {
+    if (Array.isArray(possibleValue)) return JSON.stringify(possibleValue);
+    return possibleValue;
+  }
+
+  const operator = Operators.find(op => op.value === opName);
+  return operator && operator.defaultValue ? operator.defaultValue : '';
+};
+
+export const getQueryFromUrl = (urlQuery: {
+  filter: unknown;
+  sort: unknown;
+}) => {
+  let urlFilters: string[] = [];
+  if (typeof urlQuery.filter === 'string') {
+    urlFilters = [urlQuery.filter];
+  } else if (Array.isArray(urlQuery!.filter)) {
+    urlFilters = urlQuery.filter;
+  }
+  const where = {
+    $and: urlFilters.map(filter => {
+      const parts = filter.split(';');
+      const col = parts[0];
+      const op = parts[1];
+      const value = parts[2];
+      return { [col]: { [op]: value } };
+    }),
+  };
+
+  let urlSorts: string[] = [];
+  if (typeof urlQuery.sort === 'string') {
+    urlSorts = [urlQuery.sort];
+  } else if (Array.isArray(urlQuery.sort)) {
+    urlSorts = urlQuery.sort;
+  }
+
+  const order_by = urlSorts.map(sort => {
+    const parts = sort.split(';');
+    const column = parts[0];
+    const type = parts[1];
+    const nulls = 'last';
+    return { column, type, nulls };
+  });
+
+  return { order_by, where };
 };
 
 export const isPostgresTimeoutError = (error: {